Make sure each li/event card is wrapped in a ul
