+++ front/app/components/EventCards/DateBlocks/styling.tsx
