+++ front/app/components/EventCards/EventCard.tsx
