+++ front/app/containers/EventsPage/EventsViewer/index.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 
-import { Box, media } from '@citizenlab/cl2-component-library';
+import { Box } from '@citizenlab/cl2-component-library';
 import moment from 'moment';
 import { MessageDescriptor } from 'react-intl';
 import { useSearchParams } from 'react-router-dom';
@@ -9,7 +9,7 @@ import styled from 'styled-components';
 import useEvents from 'api/events/useEvents';
 import { PublicationStatus } from 'api/projects/types';
 
-import EventCard from 'components/EventCard';
+import EventCards from 'components/EventCards';
 import Pagination from 'components/Pagination';
 
 import { updateSearchParams } from 'utils/cl-router/updateSearchParams';
@@ -21,22 +21,6 @@ import messages from '../messages';
 import EventsMessage from './EventsMessage';
 import EventsSpinner from './EventsSpinner';
 import TopBar from './TopBar';
-
-interface IStyledEventCard {
-  last: boolean;
-}
-
-const StyledEventCard = styled(EventCard)<IStyledEventCard>`
-  flex: 0 0 32.3%;
-
-  ${media.tablet`
-  flex: 0 0 48.8%;
-`}
-
-  ${media.phone`
-  flex: 0 0 100%;
-`}
-`;
 
 const StyledPagination = styled(Pagination)`
   justify-content: center;
@@ -206,17 +190,7 @@ const EventsViewer = ({
       {isLoading && <EventsSpinner />}
       {!isNilOrError(events) && (
         <>
-          <Box display="flex" flexWrap="wrap" gap="16px" as="ul" px="0px">
-            {events.data.length > 0 &&
-              events.data.map((event, i) => (
-                <StyledEventCard
-                  id={event.id}
-                  event={event}
-                  last={events.data.length - 1 === i}
-                  key={event.id}
-                />
-              ))}
-          </Box>
+          <EventCards events={events} />
 
           {events.data.length === 0 && (
             <EventsMessage message={fallbackMessage} />
